Clear the userDB and retry when we get a 412 on restore

Former-commit-id: d2687b868de956aa3cfe8aefd8fd2c676f23e7cc

+++ src/main/java/services/RestoreFactory.java
@@ -19,6 +19,7 @@ import org.javarosa.xml.util.InvalidStructureException;
 import org.javarosa.xml.util.UnfullfilledRequirementsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.http.HttpEntity;
@@ -26,6 +27,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import org.w3c.dom.Document;
 import org.w3c.dom.NamedNodeMap;
@@ -94,7 +96,7 @@ public class RestoreFactory {
     private FormplayerStorageFactory storageFactory;
 
     @Autowired
-    private RestTemplate hqRest;
+    private RestTemplateBuilder restTemplateBuilder;
 
     @Resource(name="redisTemplateLong")
     private ValueOperations<String, Long> valueOperations;
@@ -159,8 +161,20 @@ public class RestoreFactory {
         if(getSqlSandbox().getLoggedInUser() != null){
             getSQLiteDB().createDatabaseFolder();
         }
-        UserSqlSandbox sandbox = restoreUser();
-        if (shouldPurge) {
+        UserSqlSandbox sandbox;
+        try {
+             sandbox = restoreUser();
+        } catch (HttpClientErrorException e) {
+            if (e.getStatusCode().value() == 412) {
+                getSQLiteDB().deleteDatabaseFile();
+                // this line has the effect of clearing the sync token
+                // from the restore URL that's used
+                sqLiteDB = new UserDB(domain, username, asUsername);
+                return performTimedSync(shouldPurge);
+            }
+            throw e;
+        }
+        if (shouldPurge && sandbox != null) {
             SimpleTimer purgeTimer = new SimpleTimer();
             purgeTimer.start();
             FormRecordProcessorHelper.purgeCases(sandbox);
@@ -405,7 +419,7 @@ public class RestoreFactory {
         log.info("Restoring at domain: " + domain + " with url: " + restoreUrl);
         downloadRestoreTimer = categoryTimingHelper.newTimer(Constants.TimingCategories.DOWNLOAD_RESTORE);
         downloadRestoreTimer.start();
-        ResponseEntity<org.springframework.core.io.Resource> response = hqRest.exchange(
+        ResponseEntity<org.springframework.core.io.Resource> response = restTemplateBuilder.build().exchange(
                 restoreUrl,
                 HttpMethod.GET,
                 new HttpEntity<String>(headers),
